+++ apps/fountain/test/fountain_web/live/environments_form_live_test.exs
@@ -54,7 +54,7 @@ defmodule FountainWeb.EnvironmentsFormLiveTest do
         },
         "repositories" => [
           %{
-            "url" => "https://github.com/BinaryBourbon/fountain",
+            "url" => "https://github.com/managoat/fountain",
             "mount_path" => "/workspace/fountain",
             "secret_key" => "GITHUB_TOKEN"
           }
@@ -92,7 +92,7 @@ defmodule FountainWeb.EnvironmentsFormLiveTest do
         },
         "repositories" => [
           %{
-            "url" => "https://github.com/BinaryBourbon/fountain",
+            "url" => "https://github.com/managoat/fountain",
             "mount_path" => "/workspace/fountain",
             "secret_key" => "GITHUB_TOKEN"
           }